requirements.txt
Django>=4.0
google-auth>=2.22.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.94.0
numpy>=1.21.0
rasterio>=1.3.0
glob2>=0.7
pandas>=1.3.0
earthengine-api>=0.1.350  # Google Earth Engine (ee)
geemap>=0.20.0
djangorestframework==3.15.2